+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ActFramework Change Log
 
 **1.8.8**
+* Error starting app when `cacheFor.dev` set to `true` #760
+* @DateFormatPattern has no effect on `Date` field #759
 * `ResourceLoader` - load files in directory into `Map` structure #758
 * Enhance `ResourceLoader` #757
 * Missing Chinese message for some validation violations #755